+++ cmsafops/NEWS.md
@@ -1,5 +1,8 @@
 # Changelog for cmsafops package
 
+## 1.4.1
+- Bug fix: in case of just one missing value, `timcumsum` could give an error
+
 ## 1.4.0
 - New operator: `ymonmedian`
 - Improve time sorting in `cmsaf.cat`